Move semantic violations container back to left and move breakpoints back to right.
