+++ src/Components/DebugContainer/DebugContainerSemantic.js
@@ -1,7 +1,6 @@
 import React, {useEffect, useRef} from "react";
 
-// eslint-disable-next-line max-len
-import {AutomatedDebuggingContainer} from "../AutomatedDebuggingContainer/AutomatedDebuggingContainer";
+import {BreakPointContainer} from "./BreakPointContainer/BreakPointContainer";
 import {VariableStackContainer} from "./VariableStackContainer/VariableStackContainer";
 import {VerticalHandle} from "./VerticalHandle/VerticalHandle";
 
@@ -14,14 +13,14 @@ import "./DebugContainer.scss";
 export function DebugContainerSemantic () {
     const debugContainerRef = useRef();
     const variableStackRef = useRef();
-    const automatedDebuggingRef = useRef();
+    const breakPointRef = useRef();
 
     const TITLE_HEIGHT = 20;
 
     const redrawContainers = () => {
         const height = debugContainerRef.current.clientHeight;
-        const containerHeight = height - 300;
-        automatedDebuggingRef.current.style.height = 300 - TITLE_HEIGHT + "px";
+        const containerHeight = height - 150;
+        breakPointRef.current.style.height = 150 - TITLE_HEIGHT + "px";
         variableStackRef.current.style.height = containerHeight - TITLE_HEIGHT + "px";
     };
 
@@ -35,12 +34,10 @@ export function DebugContainerSemantic () {
             <div className="section" ref={variableStackRef}>
                 <VariableStackContainer />
             </div>
-            <VerticalHandle topDiv={variableStackRef} bottomDiv={automatedDebuggingRef}/>
-            <div className="w-100 title" style={{height: TITLE_HEIGHT + "px"}}>
-                Semantic Violations
-            </div>
-            <div className="section" ref={automatedDebuggingRef}>
-                <AutomatedDebuggingContainer />
+            <VerticalHandle topDiv={variableStackRef} bottomDiv={breakPointRef}/>
+            <div className="w-100 title" style={{height: TITLE_HEIGHT + "px"}}>Breakpoints</div>
+            <div className="section" ref={breakPointRef}>
+                <BreakPointContainer />
             </div>
         </div>
     );
